requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
future==0.18.2
gym==0.17.2
idna==2.9
joblib==0.15.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
plotly==4.8.1
pycolab==1.2
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==19.0.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
six==1.15.0
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
urllib3==1.25.9
visdom==0.1.8.9
websocket-client==0.57.0
